+++ lib/lor/lib/utils/utils.lua
@@ -68,7 +68,8 @@ end
 function _M.json_encode(data, empty_table_as_object)
     local json_value
     if json.encode_empty_table_as_object then
-        json.encode_empty_table_as_object(empty_table_as_object or false) -- empty table encoded as array default
+        -- empty table encoded as array default
+        json.encode_empty_table_as_object(empty_table_as_object or false) 
     end
     if require("ffi").os ~= "Windows" then
         json.encode_sparse_array(true)